Various improvements -API change  iot.basket ->iot.import -cleanup of all baskets - property derivation for attached bats

+++ sql/backends/monet5/sql_scenario.c
@@ -1187,7 +1187,9 @@ SQLCacheRemove(Client c, str nme)
 		throw(MAL, "cache.remove", "internal error, symbol missing\n");
 	if (getInstrPtr(s->def, 0)->token == FACTORYsymbol)
 		shutdownFactoryByName(c, c->nspace, nme);
-	else
-		deleteSymbol(c->nspace, s);
+	else{
+		if( s->def->inuse ==0)
+			deleteSymbol(c->nspace, s);
+	}
 	return MAL_SUCCEED;
 }
